feat: stepperInput component common component 화
